Move starting point to separate class
